backend/py_neuro/pyproject.toml
[tool.poetry]
name = "py-neuro"
version = "0.1.0"
description = "Learn through Scientifically proven methods!"
authors = ["Ishu <email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
motor = "^3.3.2"
python-multipart = "^0.0.6"
pymupdf = "^1.23.8"
python-pptx = "^0.6.22"
python-docx = "^1.1.0"
pillow = "^10.2.0"
google-generativeai = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"